fix create name unique in memberpack, delete unexist memberpack

+++ services/MemberPackService.js
@@ -12,8 +12,13 @@ const createMemberPackService = async (
     throw new Error("Invalid price or duration number");
   }
   try {
+    const checkMemberPack =
+      await connection.memberPackRepository.findMemberPackByName(name);
+    if (checkMemberPack) {
+      throw new Error("This name has been taken");
+    }
     const memberPack =
-      connection.memberPackRepository.createMemberPackRepository({
+      await connection.memberPackRepository.createMemberPackRepository({
         name,
         description,
         price,
@@ -36,7 +41,7 @@ const updateMemberPackService = async (
   const connection = new DatabaseTransaction();
   try {
     const memberPack =
-      connection.memberPackRepository.updateMemberPackRepository(
+      await connection.memberPackRepository.updateMemberPackRepository(
         id,
         name,
         description,
@@ -49,30 +54,36 @@ const updateMemberPackService = async (
     throw new Error(error.message);
   }
 };
-const getMemberPackService = (id) => {
+const getMemberPackService = async (id) => {
   const connection = new DatabaseTransaction();
   try {
     const memberPack =
-      connection.memberPackRepository.getMemberPackRepository(id);
+      await connection.memberPackRepository.getMemberPackRepository(id);
     return memberPack;
   } catch (error) {
     throw new Error(error.message);
   }
 };
-const getAllMemberPackService = () => {
+const getAllMemberPackService = async () => {
   const connection = new DatabaseTransaction();
   try {
-    const memberPacks = connection.memberPackRepository.getAllPackRepository();
+    const memberPacks =
+      await connection.memberPackRepository.getAllPackRepository();
     return memberPacks;
   } catch (error) {
     throw new Error(error.message);
   }
 };
-const deleteMemberPackService = (id) => {
+const deleteMemberPackService = async (id) => {
   const connection = new DatabaseTransaction();
   try {
+    const checkMemberPack =
+      await connection.memberPackRepository.getMemberPackRepository(id);
+    if (!checkMemberPack) {
+      throw new Error("Member Pack not found");
+    }
     const memberPack =
-      connection.memberPackRepository.deleteMemberPackRepository(id);
+      await connection.memberPackRepository.deleteMemberPackRepository(id);
     return memberPack;
   } catch (error) {
     throw new Error(error.message);
